add fetch to parte, change components to files, add modal details parte
