+++ package.json
@@ -23,7 +23,7 @@
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
-    "react-icons": "^5.0.1",
+    "react-icons": "^5.3.0",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
